Changed Reviews model to use a double instead of integer for the score attribute

+++ src/main/java/org/ani7/anirec/models/Reviews.java
@@ -6,77 +6,77 @@ import javax.persistence.*;
 @Table(name = "Reviews")
 public class Reviews {
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int reviewId;
-  private int score;
-  private String review;
-  @ManyToOne
-  @JoinColumn(name = "AnimeId")
-  private Anime anime;
-  @ManyToOne
-  @JoinColumn(name = "UserName")
-  private User user;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int reviewId;
+    private double score;
+    private String review;
+    @ManyToOne
+    @JoinColumn(name = "AnimeId")
+    private Anime anime;
+    @ManyToOne
+    @JoinColumn(name = "UserName")
+    private User user;
 
-  public Reviews() {
-  }
+    public Reviews() {
+    }
 
-  public Reviews(int reviewId, int score, String review, Anime anime, User user) {
-    this.reviewId = reviewId;
-    this.score = score;
-    this.review = review;
-    this.anime = anime;
-    this.user = user;
-  }
+    public Reviews(int reviewId, double score, String review, Anime anime, User user) {
+        this.reviewId = reviewId;
+        this.score = score;
+        this.review = review;
+        this.anime = anime;
+        this.user = user;
+    }
 
-  public Reviews(int score, String review, Anime anime, User user) {
-    this.score = score;
-    this.review = review;
-    this.anime = anime;
-    this.user = user;
-  }
+    public Reviews(double score, String review, Anime anime, User user) {
+        this.score = score;
+        this.review = review;
+        this.anime = anime;
+        this.user = user;
+    }
 
-  public Reviews(int reviewId) {
-    this.reviewId = reviewId;
-  }
+    public Reviews(int reviewId) {
+        this.reviewId = reviewId;
+    }
 
-  public int getReviewId() {
-    return reviewId;
-  }
+    public int getReviewId() {
+        return reviewId;
+    }
 
-  public void setReviewId(int reviewId) {
-    this.reviewId = reviewId;
-  }
+    public void setReviewId(int reviewId) {
+        this.reviewId = reviewId;
+    }
 
-  public int getScore() {
-    return score;
-  }
+    public double getScore() {
+        return score;
+    }
 
-  public void setScore(int score) {
-    this.score = score;
-  }
+    public void setScore(double score) {
+        this.score = score;
+    }
 
-  public String getReview() {
-    return review;
-  }
+    public String getReview() {
+        return review;
+    }
 
-  public void setReview(String review) {
-    this.review = review;
-  }
+    public void setReview(String review) {
+        this.review = review;
+    }
 
-  public Anime getAnime() {
-    return anime;
-  }
+    public Anime getAnime() {
+        return anime;
+    }
 
-  public void setAnime(Anime anime) {
-    this.anime = anime;
-  }
+    public void setAnime(Anime anime) {
+        this.anime = anime;
+    }
 
-  public User getUser() {
-    return user;
-  }
+    public User getUser() {
+        return user;
+    }
 
-  public void setUser(User user) {
-    this.user = user;
-  }
+    public void setUser(User user) {
+        this.user = user;
+    }
 }
